feat: rename keyboard.js
