+++ src/index.js
@@ -1,8 +1,8 @@
-import { Keyboard } from './keyboard';
+import { Keyboard } from './keyboardClass';
 
 window.addEventListener("DOMContentLoaded", () => {
   const keyboard = new Keyboard();
-  
+
   keyboard.generateKeyboard();
   keyboard.paintKeyboard();
 });
